refactor(ui-react): consolidate tool result previews
